Add several new class for calculation.
