+++ PrefabHouseTools/PrefabHouseTools.csproj
@@ -55,9 +55,8 @@
     <Compile Include="CmdAutoRouteE.cs" />
     <Compile Include="CmdSetCircuit.cs" />
     <Compile Include="Command.cs" />
-    <Compile Include="RoomInfo.cs" />
+    <Compile Include="CmdAutoRouteEobject.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RoomInfoElec.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="PrefabHouseTools.addin" />
